+++ test/services/printers/run_print_command_test.rb
@@ -29,4 +29,40 @@ class Printers::RunPrintCommandTest < ActiveSupport::TestCase
       Printers::RunPrintCommand.call(path: "/tmp/label.png", command: "")
     end
   end
+
+  test "inserts precut for ptouch commands when enabled" do
+    runner = lambda do |shell_command|
+      assert_equal "/usr/bin/ptouch print --precut -i #{Shellwords.escape('/tmp/label.png')} --model X", shell_command
+      [ "", "", Struct.new(:success?).new(true) ]
+    end
+
+    assert Printers::RunPrintCommand.call(
+      path: "/tmp/label.png",
+      command: "/usr/bin/ptouch print -i FILENAME --model X",
+      precut_before: true,
+      runner: runner
+    )
+  end
+
+  test "does not duplicate precut flag when already present" do
+    command = Printers::RunPrintCommand.new(
+      path: "/tmp/label.png",
+      command: "/usr/bin/ptouch print --precut -i FILENAME",
+      precut_before: true
+    )
+
+    assert_equal "/usr/bin/ptouch print --precut -i #{Shellwords.escape('/tmp/label.png')}",
+                 command.build_shell_command
+  end
+
+  test "skips precut for non-ptouch commands" do
+    command = Printers::RunPrintCommand.new(
+      path: "/tmp/label.png",
+      command: "/usr/local/bin/print-label FILENAME",
+      precut_before: true
+    )
+
+    assert_equal "/usr/local/bin/print-label #{Shellwords.escape('/tmp/label.png')}",
+                 command.build_shell_command
+  end
 end